Move exercise related code from session to exercise
- Removed exercises file, since they'll be handled a different
way later.
- Removed file file. It was no longer being used.
